pyproject.toml
[project]
name = "cairnsearch"
version = "1.0.0"
description = "Privacy-first hybrid search and local RAG over personal document collections (BM25 + dense retrieval + local LLM via Ollama)."
requires-python = ">=3.11"
dependencies = [
    # Core
    "watchdog>=3.0.0",
    
    # Document extraction
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.8.0",
    
    # API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # Config
    "tomli>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Vector operations
    "numpy>=1.24.0",
]

[project.optional-dependencies]
rag = ["httpx>=0.27.0"]
ocr = ["pytesseract>=0.3.10", "pillow>=10.0.0"]
security = ["cryptography>=41.0.0"]
all = ["httpx>=0.27.0", "pytesseract>=0.3.10", "pillow>=10.0.0", "cryptography>=41.0.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
cairnsearch = "cairnsearch.cli.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
